+++ pephub/routers/api/v1/project.py
@@ -232,15 +232,21 @@ async def delete_a_pep(
             status_code=404, detail=f"Project {namespace}/{project}:{tag} not found"
         )
 
-    agent.project.delete(namespace, project, tag=tag)
+    try:
+        agent.project.delete(namespace, project, tag=tag)
+        return JSONResponse(
+            content={
+                "message": "PEP deleted.",
+                "registry": f"{namespace}/{project}:{tag}",
+            },
+            status_code=202,
+        )
 
-    return JSONResponse(
-        content={
-            "message": "PEP deleted",
-            "registry": f"{namespace}/{project}:{tag}",
-        },
-        status_code=202,
-    )
+    except Exception as e:
+        raise HTTPException(
+            status_code=400,
+            detail=f"Could not delete PEP. Server error: {e}",
+        )
 
 
 @project.get("/samples")