+++ web/src/hooks/mutations/useConfigMutation.ts
@@ -3,6 +3,8 @@ import { AxiosError } from 'axios';
 import { toast } from 'react-hot-toast';
 
 import { editProjectConfig } from '../../api/project';
+import { extractError, extractErrorMessage } from '../../utils/etc';
+
 
 export const useConfigMutation = (
   namespace: string,
@@ -19,8 +21,13 @@ export const useConfigMutation = (
       queryClient.invalidateQueries([namespace, project, tag]);
       toast.success('Successfully updated project config');
     },
-    onError: (error: AxiosError) => {
-      toast.error(`Error updating project config: ${error}`);
+    onError: (err: AxiosError) => {
+      // extract out error message if it exists, else unknown
+      const errorMessage = extractErrorMessage(err);
+      const error = extractError(err);
+      toast.error(`${errorMessage}: ${error}`, {
+        duration: 5000,
+      });
     },
   });